# lib/api-zod/src/generated/types/streamingProvider.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StreamingProvider {
  name: string;
  logo?: string | null;
  /** flatrate, rent, buy, free, or ads */
  type: string;
  providerId: number;
  tmdbUrl: string;
  /** Direct JustWatch link for the selected country */
  watchUrl: string;
}